add random avatars and delete dialog

+++ homework-6/post-it-up/frontend/src/components/Comment/Comment.js
@@ -1,28 +1,45 @@
 import { Component } from "react";
+import { ConfirmDialog } from "../";
 import { deleteComment } from "../../api/Api";
-import styles from "./Comment.module.css";
+import { getElemAtRandomIndex } from "../../helpers/getElemAtRandomIndex";
+import { AVATARS } from "../../constants/Constant";
 import rateStar from "../../assets/star.png";
-import user from "../../assets/user.png";
 import clear from "../../assets/clear.png";
+import styles from "./Comment.module.css";
 
 class Comment extends Component {
+  state = {
+    openDialog: false,
+    avatar: getElemAtRandomIndex(AVATARS),
+  };
+
   handleDeleteComment = async (postId, commentId) => {
     const filteredComments = await deleteComment(postId, commentId);
     this.props.deleteBtnClick(filteredComments.data);
   };
 
+  handleDialogClose = () => {
+    this.setState({ openDialog: false });
+  };
+
   render() {
     const { text, rating, id } = this.props.comment;
+
     return (
       <p className={styles.comment}>
-        <img className={styles.avatar} src={user} />
+        <img className={styles.avatar} src={this.state.avatar} alt="avatar" />
         {text}
         <span className={styles.rate}>
-          <img src={rateStar} />
+          <img src={rateStar} alt="rate" />
           {rating}
         </span>
+        <ConfirmDialog
+          open={this.state.openDialog}
+          handleClose={this.handleDialogClose}
+          handleDelete={() => this.handleDeleteComment(this.props.postId, id)}
+        />
         <img
-          onClick={() => this.handleDeleteComment(this.props.postId, id)}
+          onClick={() => this.setState({ openDialog: true })}
           src={clear}
           className={styles.deleteBtn}
           alt="clear"
